feat added Clean Architecture TDD
