+++ lib/data/text_data.dart
@@ -1,0 +1,4 @@
+class DataPalindrome {
+  String text = '';
+  bool answer = false;
+}
