+++ lib/presentation/app.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:palindrom/presentation/widgets/home_page_screen.dart';
+
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: const MyHomePage(title: 'Palindrome'),
+    );
+  }
+}
